Fix fixHtmlFunc to wrap extracted JS in setTimeout call.
We do this because a JsExp's application does not happen at the point when we
evaluate that JS, so the DOM elements that the extracted JavaScript may refer
to will not be in the DOM when the expression is evaluated. Instead, we wrap
them in a setTimeout so that the rest of the expression has a chance to
actually execute against the DOM, and only after that will the extracted JS
that assumes the HTML is in the DOM execute.